requirements.txt
fastapi
uvicorn[standard]
pandas
tensorflow-cpu==2.16.1
python-multipart